Only assemble the convection terms for the system matrix in every time step.
On a stationary domain, when solving the Navier-Stokes equations with a semi-implicit convection term, only the convection-terms are time dependent, so we dont need to assemble the whole matrix in every step. In those cases, assemble C(u) in each time step, and solve the system (A+C(u))u = f.