Implement more computation modes for the boundary stress method.
The different modes are available by using the bitwise or operator on the fields in the Stress enum in FlowProblems.h.

Also let the nitche penalty parameter be a separate member value of the CutFEMProblem class, to make it easier available, and avoid setting it to different values in different places.